requirements_pip.txt
dnspython
pyyaml
-e git://github.com/fritzy/SleekXMPP.git#egg=sleekxmpp
-e git://github.com/hgrecco/SleekBot.git#egg=sleekbot

